pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopy>=2.4.0",
    "google-adk>=1.25.0",
    "python-dateutil>=2.8.2",
    "requests>=2.30.0",
    "timezonefinder>=6.1.4",
    "tzdata>=2023.3",
]
